Compiles on vogon.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@784 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Utility/GameState.cpp
@@ -1184,8 +1184,8 @@ void GameState::mouseMove(int dx, int dy, int x, int y) {
    mouseX = shipControlX;
    mouseY = shipControlY;
    
-   clientCommand.mouseX = mouseX;
-   clientCommand.mouseY = mouseY;
+   clientCommand.mouseX = (float) mouseX;
+   clientCommand.mouseY = (float) mouseY;
 
    shipControlX = shipControlX / p2wx(gameSettings->GW);
    shipControlY = shipControlY / p2wy(0);
@@ -1195,12 +1195,12 @@ void GameState::mouseMove(int dx, int dy, int x, int y) {
 
    if(!ship->flyingAI->isEnabled()) {
       if (doYaw) {
-         clientCommand.yawSpeed = -shipControlX;
+         clientCommand.yawSpeed = (float) -shipControlX;
       } else {
-         clientCommand.rollSpeed = shipControlX;
+         clientCommand.rollSpeed = (float) shipControlX;
       }
 
-      clientCommand.pitchSpeed = shipControlY;
+      clientCommand.pitchSpeed = (float) shipControlY;
    }
 }
